split monolithic element definitions/tests into many files
