+++ browsergui/__init__.py
@@ -6,8 +6,9 @@ The central concepts here are:
 - The "element," which ...
 """
 
-from . import server, gui, elements, utilities
+from . import server, gui, elements, events, utilities
 from .gui import GUI
 from .elements import Element, Container, Button, Text, Paragraph, CodeSnippet, CodeBlock, Link
 from .server import run
+from .events import *
 from .utilities import RepeatingTimer, call_in_background
