+++ browsergui/elements/__init__.py
@@ -4,10 +4,6 @@ import weakref
 import xml.dom.minidom
 import xml.parsers.expat
 
-CLICK = "click"
-KEYDOWN = "keydown"
-KEYUP = "keyup"
-
 _unique_id_counter = 0
 def unique_id():
   """Returns a new string suitable for an :class:`Element` id every time it's called."""
@@ -253,83 +249,6 @@ class Element(object):
     # Probably requires a styling system.
 
 
-class Text(Element):
-  """Some simple text."""
-  def __init__(self, text, tag_name="span"):
-    if not isinstance(text, str):
-      raise TypeError(text)
-    super(Text, self).__init__(html="<{tag}></{tag}>".format(tag=tag_name))
-    self._text = xml.dom.minidom.Text()
-    self._text.data = text
-    self.tag.appendChild(self._text)
-
-  @property
-  def text(self):
-    """docstring"""
-    return self._text.data
-  @text.setter
-  def text(self, value):
-    """docstring"""
-    if self.text == value:
-      return
-
-    self._text.data = value
-    if self.gui is not None:
-      self.gui.send_command("$({selector}).text({text})".format(selector=json.dumps("#"+self.id), text=json.dumps(self.text)))
-
-class CodeSnippet(Text):
-  """Inline text representing computer code."""
-  def __init__(self, text):
-    super(CodeSnippet, self).__init__(text, tag_name="code")
-    self.tag.attributes['style'] = 'white-space: pre;'
-class Paragraph(Text):
-  """A block of plain text."""
-  def __init__(self, text):
-    super(Paragraph, self).__init__(text, tag_name="p")
-class CodeBlock(Text):
-  """A block of computer code."""
-  def __init__(self, text):
-    super(CodeBlock, self).__init__(text, tag_name="pre")
-
-class Link(Text):
-  """A hyperlink to some URL."""
-  def __init__(self, text, url):
-    super(Link, self).__init__(text, tag_name="a")
-    self.tag.attributes['target'] = '_blank'
-    self.url = url
-
-  @property
-  def url(self):
-    return self.tag.attributes['href'].value
-  @url.setter
-  def url(self, value):
-    self.tag.attributes['href'] = value
-
-class Button(Text):
-  """A simple button that does something when clicked."""
-  def __init__(self, text="Click!", callback=None):
-    """
-    :param text: the label of the button
-    :type text: str
-    :param callback: the function to be called
-    :type callback: function of zero arguments
-    """
-    if not isinstance(text, str):
-      raise TypeError(text)
-    super(Button, self).__init__(text, tag_name="button")
-    if callback is not None:
-      self.set_callback(callback)
-
-  def set_callback(self, callback):
-    """Sets the function to be called whenever the button is clicked.
-
-    :param callback: the function to be called
-    :type callback: function of zero arguments
-    """
-    if self.callbacks[CLICK]:
-      self.remove_callback(CLICK, self.callbacks[CLICK][0])
-    self.add_callback(CLICK, (lambda event: callback()))
-
 class Container(Element):
   """Contains and groups other elements."""
   def __init__(self, *children, **kwargs):
@@ -340,3 +259,7 @@ class Container(Element):
     """
     self._inline = kwargs.pop("inline", False)
     super(Container, self).__init__(tag_name=("span" if self._inline else "div"), children=children, **kwargs)
+
+from .text import Text, Paragraph, CodeSnippet, CodeBlock
+from .button import Button
+from .link import Link
