+++ browsergui/events/__init__.py
@@ -1,0 +1,3 @@
+CLICK = "click"
+KEYDOWN = "keydown"
+KEYUP = "keyup"
